requirements.txt
# --- core data/plot ---
pandas>=2.0
numpy>=1.24
matplotlib>=3.7

# --- ml / metrics ---
scikit-learn>=1.3

# --- nlp model ---
transformers>=4.41
tokenizers>=0.15
torch>=2.1
tqdm>=4.66

# --- I/O & utils ---
openpyxl>=3.1
PyYAML>=6.0
beautifulsoup4>=4.12
requests>=2.31
